980231-part1

+++ app/src/main/java/ir/saa/android/mt/model/daos/MenuDao.java
@@ -17,7 +17,7 @@ public interface MenuDao {
     LiveData<List<Menu>> getMenus();
 
     @Insert
-    int insertMenu(Menu menu);
+    Long insertMenu(Menu menu);
 
 
     @Delete
